+++ lib/grepfruit/search_results.rb
@@ -1,6 +1,6 @@
 module Grepfruit
   class SearchResults
-    attr_reader :all_lines, :raw_matches, :total_files_with_matches
+    attr_reader :all_lines, :raw_matches, :total_files_with_matches, :match_count
     attr_accessor :total_files
 
     def initialize
@@ -8,10 +8,15 @@ module Grepfruit
       @raw_matches = []
       @total_files = 0
       @total_files_with_matches = 0
+      @match_count = 0
     end
 
     def increment_files_with_matches
       @total_files_with_matches += 1
+    end
+
+    def add_match_count(count)
+      @match_count += count
     end
 
     def add_lines(lines)